Add the first basic frontend (#11)
* - Add Angular 2 Quickstart and modify it to transpile into folder "public"

* - Add new angular version

* - Create sign in page (visually, no cleanup)

* - Fix breaking UI as of Angular Material Beta 2 update

* - Routing
- Juggling components (SignIn, PageNotFound)

* - Add a ton of components
- First simplified version of the dashboard with sidebar navigation and toolbar
- Simplified sign up and error messages

* - Final account touches
- Removed old account in the toolbar
- Added some more routing (sign in, sign out)

* - Change kmi prefix to kio

* Add Travis CI

* Fixed missed before unit tests

* Rework the AppComponent unit test

This is part of the rework of the before unused unit tests to actually
test the components.

The AppComponent is being tested for:
- Correct creation
- Containment of one router-outlet

* Remove Travis CI configuration file

With my at the time rather limited knowledge of Travis CI, I thought
that it might work to add configuration files to subdirectories. Well,
it did not. What a shame.

The Travis CI configuration for the frontend will be added by Dennis to
the overall configuration file.

* Add hammerjs to app and tests

This commit adds hammerjs in order to support gesture controls for
Angular Material.