Add basic socket service and user service
This commit adds a basic services to connect to the backend, namely
SocketService and UserService. For this matter protobufjs has been added
and the TypeScript version was bumped to ~2.2.0.

Furthermore, the protobuf packages have been changed while keeping them
intact for the go code with the option go_package.

A declaration file for the protocol.json file has also been added in
order to import it in TypeScript.